requirements.txt
tensorflow-gpu>=2.10.0
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=8.0.0
streamlit>=1.0.0
matplotlib>=3.4.0
pandas>=1.3.0
